refactored client components to adhere to linter
