+++ client/src/components/videos.jsx
@@ -1,23 +1,21 @@
 import React from 'react';
 import { PropTypes } from 'prop-types';
 
-function Videos({ video }) {
-  return (
-    <div className="carousel-component">
-      <video
-        playsInline
-        autoPlay
-        muted
-        loop
-      >
-        <source
-          src={video.url}
-          type="video/mp4"
-        />
-      </video>
-    </div>
-  );
-}
+const Videos = ({ video }) => (
+  <div className="carousel-component">
+    <video
+      playsInline
+      autoPlay
+      muted
+      loop
+    >
+      <source
+        src={video.url}
+        type="video/mp4"
+      />
+    </video>
+  </div>
+);
 
 Videos.propTypes = {
   video: PropTypes.shape({
